# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pangolin Catalog API
  description: Multi-tenant Apache Iceberg REST Catalog with advanced features including federated catalogs, credential vending, and fine-grained permissions
  contact:
    name: Pangolin API Support
    email: support@pangolin.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
...
  /api/v1/auth/revoke:
    post:
      tags:
      - Tokens
      summary: Revoke the current user's token (logout)
      operationId: revoke_current_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
        required: true
      responses:
        '200':
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeTokenResponse'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/auth/revoke/{token_id}:
    post:
      tags:
      - Tokens
      summary: Admin endpoint to revoke any token by ID
      operationId: revoke_token_by_id
      parameters:
      - name: token_id
        in: path
        description: Token ID to revoke
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
        required: true
      responses:
        '200':
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeTokenResponse'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
  /api/v1/tenants:
    get:
      tags:
      - Tenants
      operationId: list_tenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantResponse'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Tenants
      operationId: create_tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '403':
          description: Forbidden in NO_AUTH mode
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/tenants/{id}:
    get:
      tags:
      - Tenants
      operationId: get_tenant
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          description: Tenant not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    put:
      tags:
      - Tenants
      operationId: update_tenant
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
        required: true
      responses:
        '200':
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          description: Tenant not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    delete:
      tags:
      - Tenants
      operationId: delete_tenant
      parameters:
      - name: id
        in: path
        description: Tenant ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Tenant deleted
        '404':
          description: Tenant not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/tokens:
    post:
      tags:
      - Tokens
      summary: Generate a JWT token for a tenant
      description: This endpoint allows generating tokens for testing and development
      operationId: generate_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTokenRequest'
        required: true
      responses:
        '200':
          description: Token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTokenResponse'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /api/v1/users:
    get:
      tags:
      - Users
      summary: List all users (filtered by permissions)
      operationId: list_users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfo'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Users
      summary: Create a new user
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    CreateTenantRequest:
      type: object
      required:
      - name
      properties:
        admin_password:
          type: string
          nullable: true
        admin_username:
          type: string
          nullable: true
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
          nullable: true
    CreateUserRequest:
      type: object
      description: Request to create a new user
      required:
      - username
      - email
      - role
      properties:
        email:
          type: string
        password:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        tenant_id:
          type: string
          format: uuid
          nullable: true
        username:
          type: string
...
    GenerateTokenRequest:
      type: object
      required:
      - tenant_id
      properties:
        expires_in_hours:
          type: integer
          format: int64
          nullable: true
          minimum: 0
        roles:
          type: array
          items:
            type: string
          nullable: true
        tenant_id:
          type: string
        username:
          type: string
          nullable: true
    GenerateTokenResponse:
      type: object
      required:
      - token
      - expires_at
      - tenant_id
      properties:
        expires_at:
          type: string
        tenant_id:
          type: string
        token:
          type: string
...
    OAuthProvider:
      type: string
      description: OAuth provider
      enum:
      - google
      - microsoft
      - github
      - okta
...
    RevokeTokenRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true
    RevokeTokenResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
...
    TenantResponse:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
          nullable: true
...
    UserInfo:
      type: object
      description: Public user information
      required:
      - id
      - username
      - email
      - role
      properties:
        email:
          type: string
        id:
          type: string
          format: uuid
        oauth-provider:
          allOf:
          - $ref: '#/components/schemas/OAuthProvider'
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        tenant-id:
          type: string
          format: uuid
          nullable: true
        username:
          type: string
    UserRole:
      type: string
      description: User role in the system
      enum:
      - root
      - tenant-admin
      - tenant-user
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/users/login or /api/v1/tokens
...